+++ Beam/app/Section.php
@@ -8,6 +8,7 @@ class Section extends Model
 {
     protected $fillable = [
         'name',
+        'external_id',
     ];
 
     protected $hidden = [